Update game state with data returned from PATCH, seems to work but consolelog showing cells behind 1 move
I think I'm logging the cells array sometime after I make a move and override the store (state) with the latest results. So how come when I console log store it seems to always be a move behind? Do I need an async/await here? Oh actually...this might be because console logs are firing before the promise caused by .then() is returning...