Update stat item css to be able to prepare for updating span
The approach I'm planning to take is to use jquery to update the ptag with stat count, but the way the CSS was this would also overwrite the span with it. By seperating these into two entities I should be able to update just the ptag without worrying about adding logic for the span.